solve problems|1
Here's a sample C++ program that meets the requirements:

cpp
Copy code
#include <iostream>
using namespace std;

const int SIZE = 50;

void initializeArray(double* ptr)
{
    for (int i = 0; i < SIZE; i++) 
    {
        if (i < SIZE / 2) // first 25 components
            *(ptr + i) = i * i;
        else // last 25 components
            *(ptr + i) = 3 * i;
    }
}

void printArray(double* ptr)
{
    for (int i = 0; i < SIZE; i++) 
    {
        cout << *(ptr + i) << " ";
        if ((i + 1) % 10 == 0) // print 10 elements per line
            cout << endl;
    }
}

int main()
{
    double alpha[SIZE];
    double* ptr = alpha;

    initializeArray(ptr);
    printArray(ptr);

    return 0;
}
In this program, we declare an array alpha of 50 components of type double. We then initialize the array using the initializeArray function, which takes a pointer to the first element of the array as its argument. Inside the function, we use a loop to set the first 25 components of the array to the square of the index variable, and the last 25 components to three times the index variable.

We then use the printArray function to output the array, again using a pointer to the first element of the array as the argument. Inside the function, we use a loop to print each element of the array, along with a space, and check if we've printed 10 elements per line. If we have, we print a new line character to start a new line.

In the main function, we declare the array alpha and initialize a pointer ptr to point to the first element of the array. We then call the initializeArray function to fill the array with values, and the printArray function to output the array.